Intr schema, enchance object builder. fix doc.
