+++ docs/fork-example.md
@@ -12,13 +12,16 @@ use Formapro\Pvm\CallbackBehavior;
 use Formapro\Pvm\ProcessEngine;
 use Formapro\Pvm\Process;
 use Formapro\Pvm\Token;
+use Formapro\Pvm\ObjectBuilderHook;
+
+(new ObjectBuilderHook())->register();
 
 $registry = new DefaultBehaviorRegistry();
 $registry->register('print_label', new CallbackBehavior(function(Token $token) {
     echo $token->getTransition()->getTo()->getLabel().' ';
 }));
 
-$process = new Process();
+$process = Process::create();
 $foo = $process->createNode();
 $foo->setLabel('foo');
 $foo->setBehavior('print_label');
